+++ src/main/java/com/drunkencod/mobtalismans/item/ModItems.java
@@ -29,6 +29,9 @@ public class ModItems {
 
     // #region Talismans:
 
+    public static final DeferredItem<BatTalismanItem> BAT_TALISMAN = ITEMS.register(
+            BatTalismanItem.REGISTRY_NAME, BatTalismanItem::new);
+
     public static final DeferredItem<CreeperTalismanItem> CREEPER_TALISMAN = ITEMS.register(
             CreeperTalismanItem.REGISTRY_NAME, CreeperTalismanItem::new);